Split source for show and sf
