+++ src/showfunctions.c
@@ -38,9 +38,9 @@
 #include <regex.h>
 #include "common.h"
 #include "config.h"
-#include "functions.h"
-#include "views.h"
-#include "menus.h"
+#include "showfunctions.h"
+#include "showviews.h"
+#include "showmenus.h"
 #include "colors.h"
 #include "show.h"
